macOS/MoltenVK: negotiate device features instead of hardcoding
Query physical device features before requesting them; treat shader_int_64
and shader_int_16 as required, everything else (shader_float_64, geometry/
tessellation shaders, fill_mode_non_solid, wide_lines, large_points) as
optional and enable only what the device reports. Record enabled optionals in
ENABLED_OPTIONAL_FEATURES; add has_device_feature() so the SPIR-V compiler
target can gate supports_fp64 on shader_float_64. De-widen abs(ComplexF32) to
stay Float32-only on devices without shader_float_64. Fixes device init on
MoltenVK/Apple Silicon (was VK_ERROR_FEATURE_NOT_PRESENT).

+++ src/device/math.jl
@@ -300,9 +300,9 @@ end
 # ══════════════════════════════════════════════════════════════════════
 
 @lava_device_override @inline function Base.abs(z::ComplexF32)
-    re = Float64(real(z))
-    im = Float64(imag(z))
-    Float32(sqrt(muladd(re, re, im * im)))
+    re = real(z)
+    im = imag(z)
+    sqrt(muladd(re, re, im * im))
 end
 
 @lava_device_override @inline function Base.abs(z::ComplexF64)
